+++ entier.h
@@ -19,7 +19,7 @@ public:
     Constante& operator-(const Constante& c)const;
     Constante& operator*(const Constante& c)const;
     Constante& operator/(const Constante& c)const;
-    Constante& operator-(int)const;
+    Constante& operator-()const;
     Constante& operator^(const Constante& c)const;
     Constante& sinus()const;
     Constante& cosinus()const;
